Fix default repo not updating
The intended use of ${{ inputs.repo || env.GITHUB_REPOSITORY }} is to
allow the variable to default to the local repository string if it is
empty.
However, when input repo is left empty, the value is saved as "" which
will prevent the value to default to the current repository.

Let's use default field to set the default field of the repo input.